+++ packages/common-cosmos/src/project/project.spec.ts
@@ -46,9 +46,9 @@ describe('project.yaml', () => {
     ) as any;
     expect(() => parseCosmosProjectManifest(cosmosManifest)).toThrow('failed to parse project.yaml');
   });
-  it('Ensure chainTypes existence on manifest deployment', () => {
+  it('Ensure chaintypes existence on manifest deployment', () => {
     const cosmosManifest = loadFromJsonOrYaml(path.join(projectsDir, './protoTest1', 'project.yaml')) as any;
     const manifest = parseCosmosProjectManifest(cosmosManifest);
-    expect(manifest.asImpl.network.chainTypes.size).toBeGreaterThan(0);
+    expect(manifest.asImpl.network.chaintypes.size).toBeGreaterThan(0);
   });
 });